+++ src/arcade/potts/sim/PottsSeries.java
@@ -259,7 +259,7 @@ public final class PottsSeries extends Series {
             // Get list of links, if valid.
             MiniBox links = box.filterBoxByTag("LINK").getIdValForTagAtt("LINK", "weight");
             for (String link : links.getKeys()) {
-                population.put("(LINK)" + TAG_SEPARATOR + link, links.getInt(link));
+                population.put("(LINK)" + TAG_SEPARATOR + link, links.getDouble(link));
             }
 
             // Get list of regions, if valid.
